requirements.txt
numpy==1.21.0
opencv_contrib_python==4.4.0.46
tensorflow==2.11.0
pandas==1.1.4
streamlit==1.15.1
pillow==9.0.0
h5py==2.10.0